+++ openosint/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "2.19.1"
+__version__ = "2.20.0"
 version = __version__